restructure test directory. utils.py got too crowded
