+++ test/unit/logic/backup/test_backup_preperator.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from subprocess import Popen
-from test.integration.logic.backup.utils import prepare_source_sink_dirs, temp_source_sink_dirs
+from test.utils.backup_test_environment import prepare_source_sink_dirs
 from typing import Generator, Optional, Tuple
 
 import pytest
